+++ src/app/components/balances/balance-pie-chart/balance-pie-chart.component.ts
@@ -1,6 +1,8 @@
 import { CurrencyPipe } from '@angular/common'
 import { Component, inject } from '@angular/core'
-import { ChartService } from '@services/chart.service'
+import { ChartBalance } from '@interfaces/chart-balance.interface'
+import { FinanceService } from '@services/finance.service'
+import { colorPalettes, pieOptions } from '@utilities/constants'
 import { BaseChartDirective } from 'ng2-charts'
 
 @Component({
@@ -9,5 +11,55 @@ import { BaseChartDirective } from 'ng2-charts'
   imports: [BaseChartDirective, CurrencyPipe],
 })
 export class BalancePieChartComponent {
-  chartService = inject(ChartService)
+  private financeService = inject(FinanceService)
+
+  chartBalance: ChartBalance = {
+    chartType: 'doughnut',
+    options: pieOptions,
+    data: {
+      labels: [],
+      datasets: [
+        {
+          data: [],
+          ...colorPalettes.balances,
+        },
+      ],
+    },
+    total: 0,
+  }
+
+  constructor() {
+    this.financeService.events.subscribe((event) => {
+      if (event.resource === 'budget' || event.resource === 'balance') {
+        this.setChartBalance()
+      }
+    })
+  }
+
+  private setChartBalance() {
+    if (!this.financeService.budget?.balances) {
+      return
+    }
+
+    let total = 0
+    const data: number[] = []
+    const labels: string[] = []
+
+    this.financeService.budget.balances().forEach((balance) => {
+      data.push(balance.amount)
+      labels.push(balance.description)
+      total += balance.amount
+    })
+    this.chartBalance.data = {
+      ...this.chartBalance.data,
+      labels,
+      datasets: [
+        {
+          ...this.chartBalance.data.datasets[0],
+          data,
+        },
+      ],
+    }
+    this.chartBalance.total = total
+  }
 }
