+++ src/app/components/revenues/revenue-pie-chart/revenue-pie-chart.component.ts
@@ -1,6 +1,8 @@
 import { CurrencyPipe } from '@angular/common'
 import { Component, inject } from '@angular/core'
-import { ChartService } from '@services/chart.service'
+import { ChartRevenue } from '@interfaces/chart-revenue.interface'
+import { FinanceService } from '@services/finance.service'
+import { colorPalettes, pieOptions } from '@utilities/constants'
 import { BaseChartDirective } from 'ng2-charts'
 
 @Component({
@@ -9,5 +11,55 @@ import { BaseChartDirective } from 'ng2-charts'
   imports: [BaseChartDirective, CurrencyPipe],
 })
 export class RevenuePieChartComponent {
-  chartService = inject(ChartService)
+  private financeService = inject(FinanceService)
+
+  chartRevenue: ChartRevenue = {
+    chartType: 'doughnut',
+    options: pieOptions,
+    data: {
+      labels: [],
+      datasets: [
+        {
+          data: [],
+          ...colorPalettes.revenues,
+        },
+      ],
+    },
+    total: 0,
+  }
+
+  constructor() {
+    this.financeService.events.subscribe((event) => {
+      if (event.resource === 'budget' || event.resource === 'revenue') {
+        this.setChartRevenue()
+      }
+    })
+  }
+
+  private setChartRevenue() {
+    if (!this.financeService.budget?.revenues) {
+      return
+    }
+
+    let total = 0
+    const data: number[] = []
+    const labels: string[] = []
+
+    this.financeService.budget.revenues().forEach((revenue) => {
+      data.push(revenue.yearlyAmount)
+      labels.push(revenue.description)
+      total += revenue.yearlyAmount
+    })
+    this.chartRevenue.data = {
+      ...this.chartRevenue.data,
+      labels,
+      datasets: [
+        {
+          ...this.chartRevenue.data.datasets[0],
+          data,
+        },
+      ],
+    }
+    this.chartRevenue.total = total
+  }
 }
